Crate assignment2

Assignment 2

This is a CI server that listens for webhooks from GitHub and runs a CI pipeline on the repository.

Modules

ci

This module implements a simple Continuous Integration (CI) system. It provides functionality to build and test local repositories.

github

A module for handling everything related to GitHub.

routes

This module contains all the routes for the application.

Constants

DOC_PATH 🔒

LOGS_PATH 🔒

The path to the directory where the logs are stored.

REPO_PATH 🔒

The path to the directory where the repositories are cloned.

SERVER_URL 🔒

The URL and port of the server.

Functions

main 🔒

Starts the CI server and handles all the routes.